Get userId only from meteor methods

+++ imports/api/comments/server/securedMethods.js
@@ -4,11 +4,10 @@ import CommentService from '/imports/api/comments/commentService';
 
 Meteor.methods({
     'secured.comment_create' (comment) {
-        return CommentService.createComment(comment);
+        return CommentService.createComment({userId: this.userId, ...comment});
     },
 
     'secured.comment_remove' (_id) {
-        CommentService.removeComment(_id);
+        CommentService.removeComment({_id, userId: this.userId});
     }
-
 });
